+++ .prettierrc
@@ -1,0 +1,5 @@
+{
+	"singleQuote": true,
+	"trailingComma": "none",
+	"semi": false
+}
