+++ package.json
@@ -17,10 +17,11 @@
 		"@typescript-eslint/parser": "5.29.0",
 		"builtin-modules": "3.3.0",
 		"esbuild": "0.17.3",
+		"eslint": "^8.50.0",
 		"obsidian": "latest",
+		"obsidian-dataview": "^0.5.58",
+		"prettier": "^3.0.3",
 		"tslib": "2.4.0",
 		"typescript": "4.7.4"
-	},
-	"dependencies": {
 	}
 }
