+++ src/main.ts
@@ -1,4 +1,3 @@
-
 import { ChatStreamPlugin } from './ChatStreamPlugin'
 
 export default ChatStreamPlugin
